+++ app/upgrades/v23/constants.go
@@ -1,7 +1,7 @@
 package v23
 
 import (
-	ibcwasmtypes "github.com/cosmos/ibc-go/modules/light-clients/08-wasm/types"
+	ibcwasmtypes "github.com/cosmos/ibc-go/modules/light-clients/08-wasm/v10/types"
 
 	"cosmossdk.io/store/types"
 
@@ -10,8 +10,12 @@ import (
 
 const (
 	// UpgradeName defines the on-chain upgrade name.
-	UpgradeName    = "v23"
-	IbcFeeStoreKey = "feeibc"
+	UpgradeName = "v23"
+	// RCUpgradeName defines the on-chain upgrade name specifically for the testnet RC upgrade.
+	RCUpgradeName           = "23.0.0-rc3"
+	IbcFeeStoreKey          = "feeibc"
+	ClientUploaderAddress   = "cosmos1raa4kyx5ypz75qqk3566c6slx2mw3qzs5ps5du"
+	IBCWasmStoreCodeTypeURL = "/ibc.lightclients.wasm.v1.MsgStoreCode"
 )
 
 var Upgrade = upgrades.Upgrade{
@@ -25,5 +29,15 @@ var Upgrade = upgrades.Upgrade{
 		Deleted: []string{
 			IbcFeeStoreKey,
 		},
+	},
+}
+
+var RCUpgrade = upgrades.Upgrade{
+	UpgradeName:          RCUpgradeName,
+	CreateUpgradeHandler: CreateRCUpgradeHandler,
+	StoreUpgrades: types.StoreUpgrades{
+		Added:   nil,
+		Renamed: nil,
+		Deleted: nil,
 	},
 }